Fix a bug in the mke2fs example.
If a filesystem had more than one block for block group descriptors (which
means that there were more than 128 block groups with 4096 byte blocks), then
the the descriptors for groups 128 and up would be wrong (just mirrors of those
for groups 0-127). You'd need a pretty big SD card to make this happen (>16GiB).