Make the cdrom code aware of the g1ata stuff.
This involved making sure that g1_ata_* locks the same mutex as the cdrom stuff
and making sure the cdrom stuff sets up the device select register properly
before trying to access the cdrom (since the BIOS syscalls don't do it).

There's still a bit of cleanup related to all of this to do in g1_ata_*, but it
works at this point, so I'm committing it now.