Add getaddrinfo(), freeaddrinfo(), gethostbyname(), and gethosbyname2() functions.
getaddrinfo() doesn't handle most of the flags available for it or service lookups by name, but it should be fine for its most common use (DNS lookups).
gethostbyname() should be avoided, in general (it isn't thread safe, as it isn't required by POSIX to be), but was added for potential compatibility with code that uses it.
gethostbyname2() is a GNU extension of gethostbyname() that allows you to specify the address family of addresses to be returned. Once again, it should be avoided for thread-safety concerns.