Remove the ENABLE_WRITES macro from flashrom.c.
I can't think of any good reason to keep this around, as it should be pretty obvious that a function called flashrom_write() or flashrom_delete() is going to write to the flashrom. So, the macro has been removed and the default state is now just as if the macro was defined.